scx_layered: Use u64 for dsq_id
Layer DSQ IDs are 64-bit, this was part of the commit
347da2af8acf ("scx_layered: Fix race condition between picking an idle CPU and the CPU going idle")
as a stray fix, resurrect it now that we've reverted the change.

Suggested-by: Tejun Heo <tj@kernel.org>

+++ scheds/rust/scx_layered/src/bpf/main.bpf.c
@@ -1680,7 +1680,7 @@ static bool keep_running(struct cpu_ctx *cpuc, struct task_struct *p,
 		 * have tasks waiting, keep running it. If there are multiple
 		 * competing preempting layers, this won't work well.
 		 */
-		u32 dsq_id = layer_dsq_id(layer->id, cpuc->llc_id);
+		u64 dsq_id = layer_dsq_id(layer->id, cpuc->llc_id);
 		if (!scx_bpf_dsq_nr_queued(dsq_id)) {
 			p->scx.slice = layer->slice_ns;
 			lstat_inc(LSTAT_KEEP, layer, cpuc);
